afex-package {afex}	R Documentation

Analysis of Factorial Experiments Maintainer: NA

Description

Convenience functions for analyzing factorial experiments using ANOVA or mixed models. aov_ez(), aov_car(), and aov_4() allow specification of between, within (i.e., repeated-measures), or mixed (i.e., split-plot) ANOVAs for data in long format (i.e., one observation per row), automatically aggregating multiple observations per individual and cell of the design. mixed() fits mixed models using lme4::lmer() and computes p-values for all fixed effects using either Kenward-Roger or Satterthwaite approximation for degrees of freedom (LMM only), parametric bootstrap (LMMs and GLMMs), or likelihood ratio tests (LMMs and GLMMs). afex_plot() provides a high-level interface for interaction or one-way plots using ggplot2, combining raw data and model estimates. afex uses type 3 sums of squares as default (imitating commercial statistical software).

Details

The DESCRIPTION file:

Package: 	 afex

Type: 	 Package

Title: 	 Analysis of Factorial Experiments

Depends: 	 R (>= 3.5.0), lme4 (>= 1.1-8)

Suggests: 	 emmeans (>= 1.4), coin, xtable, parallel, plyr, optimx, nloptr, knitr, rmarkdown, R.rsp, lattice, latticeExtra, multcomp, testthat, mlmRev, dplyr, tidyr, dfoptim, Matrix, psychTools, ggplot2, MEMSS, effects, carData, ggbeeswarm, nlme, cowplot, jtools, ggpubr, ggpol, MASS, glmmTMB, brms, rstanarm, statmod

Imports: 	 pbkrtest (>= 0.4-1), lmerTest (>= 3.0-0), car, reshape2, stats, methods, utils

Description: 	 Convenience functions for analyzing factorial experiments using ANOVA or mixed models. aov_ez(), aov_car(), and aov_4() allow specification of between, within (i.e., repeated-measures), or mixed (i.e., split-plot) ANOVAs for data in long format (i.e., one observation per row), automatically aggregating multiple observations per individual and cell of the design. mixed() fits mixed models using lme4::lmer() and computes p-values for all fixed effects using either Kenward-Roger or Satterthwaite approximation for degrees of freedom (LMM only), parametric bootstrap (LMMs and GLMMs), or likelihood ratio tests (LMMs and GLMMs). afex_plot() provides a high-level interface for interaction or one-way plots using ggplot2, combining raw data and model estimates. afex uses type 3 sums of squares as default (imitating commercial statistical software).

URL: 	 http://afex.singmann.science/, https://github.com/singmann/afex

BugReports: 	 https://github.com/singmann/afex/issues

License: 	 GPL (>=2)

Encoding: 	 UTF-8

VignetteBuilder: 	 knitr, R.rsp

Authors@R: 	 c(person(given="Henrik", family="Singmann", role=c("aut", "cre"), email="singmann@gmail.com", comment=c(ORCID="0000-0002-4842-3657")), person(given="Ben", family="Bolker", role=c("aut")), person(given="Jake",family="Westfall", role=c("aut")), person(given="Frederik", family="Aust", role=c("aut"), comment = c(ORCID = "0000-0003-4900-788X")), person(given="Mattan S.",family="Ben-Shachar", role=c("aut")), person(given="Soren", family="Hojsgaard", role=c("ctb")), person(given="John", family="Fox", role=c("ctb")), person(given="Michael A.", family="Lawrence", role=c("ctb")), person(given="Ulf", family="Mertens", role=c("ctb")), person(given="Jonathon", family="Love", role=c("ctb")), person(given="Russell", family="Lenth", role=c("ctb")), person(given="Rune", family="Haubo Bojesen Christensen", role=c("ctb")))

Version: 	 0.28

RoxygenNote: 	 7.1.0

LazyData: 	 true

RemoteType: 	 github

RemoteHost: 	 api.github.com

RemoteRepo: 	 afex

RemoteUsername: 	 singmann

RemoteRef: 	 master

RemoteSha: 	 71e22f0020399de1b35189d7c0dd4e5a2729b843

GithubRepo: 	 afex

GithubUsername: 	 singmann

GithubRef: 	 master

GithubSHA1: 	 71e22f0020399de1b35189d7c0dd4e5a2729b843

Author: 	 Henrik Singmann [aut, cre] (<https://orcid.org/0000-0002-4842-3657>), Ben Bolker [aut], Jake Westfall [aut], Frederik Aust [aut] (<https://orcid.org/0000-0003-4900-788X>), Mattan S. Ben-Shachar [aut], Soren Hojsgaard [ctb], John Fox [ctb], Michael A. Lawrence [ctb], Ulf Mertens [ctb], Jonathon Love [ctb], Russell Lenth [ctb], Rune Haubo Bojesen Christensen [ctb]

Maintainer: 	 Henrik Singmann <singmann@gmail.com>

Author(s)

Maintainer: Henrik Singmann singmann@gmail.com (ORCID)

Authors:

Ben Bolker

Jake Westfall

Frederik Aust (ORCID)

Mattan S. Ben-Shachar

Other contributors:

Søren Højsgaard [contributor]

John Fox [contributor]

Michael A. Lawrence [contributor]

Ulf Mertens [contributor]

Jonathon Love [contributor]

Russell Lenth [contributor]

Rune Haubo Bojesen Christensen [contributor]

See Also

Useful links:

http://afex.singmann.science/

https://github.com/singmann/afex

Report bugs at https://github.com/singmann/afex/issues

[Package afex version 0.28 Index]