aov.car {afex}	R Documentation

Deprecated functions

Description

These functions have been renamed and deprecated in afex: aov.car() (use aov_car()), ez.glm() (use aov_ez()), aov4() (use aov_4()).

Usage


aov.car(...)

ez.glm(...)

aov4(...)

Arguments

...
arguments passed from the old functions of the style foo.bar() to the new functions foo_bar()

[Package afex version 0.28 Index]