fhch2010 {afex}	R Documentation

Data from Freeman, Heathcote, Chalmers, & Hockley (2010)

Description

Lexical decision and word naming latencies for 300 words and 300 nonwords presented in Freeman, Heathcote, Chalmers, and Hockley (2010). The study had one between-subjects factors, "task" with two levels ("naming" or "lexdec"), and four within-subjects factors: "stimulus" type with two levels ("word" or "nonword"), word "density" and word "frequency" each with two levels ("low" and "high") and stimulus "length" with three levels (4, 5, and 6).

Usage


fhch2010

Format

A data.frame with 13,222 obs. of 9 variables:

id

participant id, factor

task

factor with two levels indicating which task was performed: "naming" or "lexdec"

stimulus

factor indicating whether the shown stimulus was a "word" or "nonword"

density

factor indicating the neighborhood density of presented items with two levels: "low" and "high". Density is defined as the number of words that differ from a base word by one letter or phoneme.

frequency

factor indicating the word frequency of presented items with two levels: "low" (i.e., words that occur less often in natural language) and "high" (i.e., words that occur more often in natural language).

length

factor with 3 levels (4, 5, or 6) indicating the number of characters of presented stimuli.

item

factor with 600 levels: 300 words and 300 nonwords

rt

response time in seconds

log_rt

natural logarithm of response time in seconds

correct

boolean indicating whether or not the response in the lexical decision task was correct or incorrect (incorrect responses of the naming task are not part of the data).

Details

In the lexical-decision condition (N = 25), subjects indicated whether each item was a word or a nonword, by pressing either the left (labeled word) or right (labeled nonword) outermost button on a 6-button response pad. The next study item appeared immediately after the lexical decision response was given. In the naming condition (N = 20), subjects were asked to name each item aloud, and items remained on screen for 3 s. Naming time was recorded by a voice key.

Items consisted of 300 words, 75 in each set making up a factorial combination of high and low density and frequency, and 300 nonwords, with equal numbers of 4, 5, and 6 letter items in each set.

Source

Freeman, E., Heathcote, A., Chalmers, K., & Hockley, W. (2010). Item effects in recognition memory for words. Journal of Memory and Language, 62(1), 1-18. http://doi.org/10.1016/j.jml.2009.09.004

Examples



data("fhch2010")
str(fhch2010)

a1 <- aov_ez("id", "log_rt", fhch2010, between = "task", 
             within = c("density", "frequency", "length", "stimulus"))
nice(a1)

if (requireNamespace("emmeans") && requireNamespace("ggplot2")) {
  afex_plot(a1, "length", "frequency", c("task", "stimulus"), error = "within")
  
  afex_plot(a1, "density", "frequency", c("task", "stimulus"), error = "within")
}


## Not run: 
a2 <- aov_ez("id", "rt", fhch2010, between = "task", 
             within = c("density", "frequency", "length", "stimulus"))
nice(a2)

if (requireNamespace("emmeans") && requireNamespace("ggplot2")) {
  afex_plot(a2, "length", "frequency", c("task", "stimulus"), error = "within")
  
  afex_plot(a2, "density", "frequency", c("task", "stimulus"), error = "within")
}

## End(Not run)

[Package afex version 0.28-0 Index]