residuals.afex_aov {afex}	R Documentation

Extract Residuals and Fitted Values from afex_aov objects

Description

Extract Residuals and Fitted Values from afex_aov objects.

Usage


## S3 method for class 'afex_aov'
residuals(object, append = FALSE, colname_residuals = ".residuals", ...)

## S3 method for class 'afex_aov'
fitted(object, append = FALSE, colname_fitted = ".fitted", ...)

Arguments

object
afex_aov object.

append
If set to TRUE returns the residuals/fitted values appended as an additional column to the long data. Recomended when data was aggragated across within conditions.

colname_residuals, colname_fitted
Name of the appended column when append = TRUE.

...
Additional arguments passed to residuals.lm/fitted.lm.

Value

A vector of residualts/fitted values corresponding to the data in object$data$long, or if append = TRUE a data frame with an additional column of residualts/fitted values.

Author(s)

Mattan S. Ben-Shachar

Examples


### Setup ANOVAs
data(obk.long, package = "afex")
between <- aov_car(value ~ treatment*gender + Error(id), data = obk.long)
within <- aov_car(value ~ 1 + Error(id/(phase*hour)), data = obk.long)
mixed <- aov_car(value ~ treatment * gender + Error(id/(phase*hour)), data = obk.long)

# All residuals call produce the warning that the data was changed during calculation.
residuals(within)
residuals(mixed)
residuals(between)

## Get residuals plus data used for fitting: 
residuals(within, append = TRUE)
residuals(mixed, append = TRUE)
residuals(between, append = TRUE)

## in case data is correctly ordered before fitting, this warning is not shown:
obk2 <- aggregate(value ~ gender + treatment + id , data = obk.long, FUN = mean)
between2 <- aov_car(value ~ treatment*gender + Error(id), data = obk2)

residuals(between2) ## no warning because
all.equal(obk2, between2$data$long[,colnames(obk2)]) ## TRUE

## Therefore okay:
obk2$residuals <- residuals(between2)


## Same for fitted values:
fitted(within)
fitted(mixed)
fitted(between)

## Get fitted values plus data used for fitting: 
fitted(within, append = TRUE)
fitted(mixed, append = TRUE)
fitted(between, append = TRUE)

[Package afex version 0.28 Index]