set_sum_contrasts {afex}	R Documentation

Set global contrasts

Description

These functions are simple wrappers to set contrasts globally via options(contrasts = ...).

Usage


set_sum_contrasts()

set_deviation_contrasts()

set_effects_contrasts()

set_default_contrasts()

set_treatment_contrasts()

Details

set_deviation_contrasts and set_effects_contrasts are wrappers for set_sum_contrasts. Likewise, set_default_contrasts is a wrapper to set_treatment_contrasts().

Value

nothing. These functions are called for their side effects to change the global options.

[Package afex version 0.28 Index]